requirement.txt
Django==1.6.4
South==0.8.4
django-braces==1.4.0
django-mptt==0.6.0
six==1.6.1
